src/agent.py
"""G1 Agent - Advanced AI agent with reasoning capabilities."""

import os
from typing import Optional, Dict, Any
from dotenv import load_dotenv
from loguru import logger

# Load environment variables
load_dotenv()


class G1Agent:
    """
    G1 Agent implementation with configurable LLM backends.
    
    Attributes:
        model: The LLM model to use
        temperature: Sampling temperature for generation
        max_tokens: Maximum tokens in response
    """
    
    def __init__(
        self,
        model: Optional[str] = None,
        temperature: Optional[float] = None,
        max_tokens: Optional[int] = None,
    ):
        """
        Initialize the G1 Agent.
        
        Args:
            model: Model name (defaults to env AGENT_MODEL or 'gpt-4')
            temperature: Temperature for generation (defaults to env AGENT_TEMPERATURE or 0.7)
            max_tokens: Max tokens in response (defaults to env AGENT_MAX_TOKENS or 2000)
        """
        self.model = model or os.getenv("AGENT_MODEL", "gpt-4")
        self.temperature = temperature or float(os.getenv("AGENT_TEMPERATURE", "0.7"))
        self.max_tokens = max_tokens or int(os.getenv("AGENT_MAX_TOKENS", "2000"))
        
        # Configure logging
        log_level = os.getenv("LOG_LEVEL", "INFO")
        logger.add("logs/agent.log", rotation="10 MB", level=log_level)
        
        logger.info(f"G1 Agent initialized with model={self.model}")
    
    def run(self, prompt: str, context: Optional[Dict[str, Any]] = None) -> str:
        """
        Run the agent with the given prompt.
        
        Args:
            prompt: The input prompt for the agent
            context: Optional context dictionary for additional information
            
        Returns:
            The agent's response as a string
        """
        logger.info(f"Running agent with prompt: {prompt[:100]}...")
        
        try:
            # TODO: Implement actual LLM integration
            response = self._generate_response(prompt, context)
            logger.info("Agent response generated successfully")
            return response
        except Exception as e:
            logger.error(f"Error during agent execution: {e}")
            raise
    
    def _generate_response(self, prompt: str, context: Optional[Dict[str, Any]] = None) -> str:
        """
        Generate a response using the configured LLM.
        
        Args:
            prompt: The input prompt
            context: Optional context dictionary
            
        Returns:
            Generated response
        """
        # Placeholder implementation
        # TODO: Integrate with OpenAI/Anthropic APIs based on model selection
        return f"Response to: {prompt}"
    
    def add_tool(self, tool_name: str, tool_function: callable) -> None:
        """
        Add a tool that the agent can use.
        
        Args:
            tool_name: Name of the tool
            tool_function: The callable tool function
        """
        # TODO: Implement tool management
        logger.info(f"Tool '{tool_name}' registered")
        pass


if __name__ == "__main__":
    # Example usage
    agent = G1Agent()
    result = agent.run("What is the meaning of life?")
    print(result)
