requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
python-dotenv>=1.0.0

# Data processing
pydantic>=2.0.0
requests>=2.31.0

# Logging and utilities
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
